refactoring: rename file to myModel
